Crate spl_transfer_hook_interface

Crate defining an interface for performing a hook on transfer, where the token program calls into a separate program with additional accounts after all other logic, to be sure that a transfer has accomplished all required preconditions.

Re-exports

pub use solana_program;

Modules

error

Error types

instruction

Instruction types

offchain

Offchain helper for fetching required accounts to build instructions

onchain

On-chain program invoke helper to perform on-chain execute with correct accounts

Constants

NAMESPACE

Namespace for all programs implementing transfer-hook

Functions

collect_extra_account_metas_seeds

Function used by programs implementing the interface, when creating the PDA, to get all of the PDA seeds

collect_extra_account_metas_signer_seeds

Function used by programs implementing the interface, when creating the PDA, to sign for the PDA

get_extra_account_metas_address

Get the state address PDA

get_extra_account_metas_address_and_bump_seed

Function used by programs implementing the interface, when creating the PDA, to also get the bump seed