spl_transfer_hook_interface

Module instruction

Instruction types

Structs

ExecuteInstruction

TLV instruction type only used to define the discriminator. The actual data is entirely managed by ExtraAccountMetaList, and it is the only data contained by this type.

InitializeExtraAccountMetaListInstruction

TLV instruction type used to initialize extra account metas for the transfer hook

UpdateExtraAccountMetaListInstruction

TLV instruction type used to update extra account metas for the transfer hook

Enums

TransferHookInstruction

Instructions supported by the transfer hook interface.

Functions

execute

Creates an Execute instruction, without the additional accounts

execute_with_extra_account_metas

Creates an Execute instruction, provided all of the additional required account metas

initialize_extra_account_meta_list

Creates a InitializeExtraAccountMetaList instruction.

update_extra_account_meta_list

Creates a UpdateExtraAccountMetaList instruction.