spl_transfer_hook_interface
Module offchain

Offchain helper for fetching required accounts to build instructions

Functions

add_extra_account_metas_for_execute

Offchain helper to get all additional required account metas for an execute instruction, based on a validation state account.

Type Aliases

AccountDataResult

Type representing the output of an account fetching function, for easy chaining between APIs

AccountFetchError

Generic error type that can come out of any client while fetching account data